+++ Drift/UI/Conversation List View/ConversationListViewController.swift
@@ -70,10 +70,6 @@ class ConversationListViewController: UIViewController {
             present(unableToAuthAlert, animated: true)
         }
         
-        if #available(iOS 11.0, *) {
-            tableView.contentInsetAdjustmentBehavior = .never
-        }
-        
         setupEmptyState()
         tableView.tableFooterView = UIView(frame: CGRect.zero)
         tableView.delegate = self
